+++ src/test/v2.0/advanced/fdc3.getInfo.ts
@@ -1,6 +1,6 @@
 import { assert, expect } from "chai";
 import { failOnTimeout, wrapPromise } from "../../../utils";
-import { closeMockAppWindow } from "../utils_2_0";
+import { closeMockAppWindow } from "../fdc3-2_0-utils";
 import { ImplementationMetadata } from "fdc3_2_0";
 import { MetadataValidator, MetadataContext, MetadataFdc3Api } from "./metadata-support-2.0";
 import { APIDocumentation2_0 } from "../apiDocuments-2.0";
