+++ src/test/v2.0/fdc3-2_0-utils.ts
@@ -1,5 +1,5 @@
 import { Channel, Context, DesktopAgent, Listener } from "fdc3_2_0";
-import { AppControlContext } from "../../common-types";
+import { AppControlContext } from "../../context-types";
 import constants from "../../constants";
 import { wait } from "../../utils";
